Refactor out excessive borrowed value optimizations in macros (#1610)
### What
Replace borrowed references with owned values in `impl_pub_methods` and
`Fn`, which are internal functions and types in the soroban-sdk-macros.

  ### Why
The excessive use of borrowed references and lifetime annotations was
adding unnecessary complexity without meaningful performance benefits in
macro expansion code that runs at compile time. Several times when I've
worked with this code I have found it harder to evolve because of the
limitation on data having to be owned externally to these functions.