Store test name and env number in env itself (#1623)
### What
Store test name and environment number directly in EnvTestState instead
of tracking them at the point where test snapshots are written.

  ### Why
Snapshot file numbering is confusing and difficult to follow when
multiple Env instances existed per test. It's because the order of the
env numbers is based on the order of the drops, not the order of the
envs being defined. This is difficult to reason about, and doesn't work
with new features I want to add where we'll write and read from other
types of files for each Env in a test.

This is technically a breaking change, but multiple env's in tests is
super rare and is supported as an edge case to ensure things still keep
working. I think it's okay to "fix this" with a break given the rareness
of its use. Anyone using it who does see a break will see there test
files flip. 1 -> 2, 2 -> 1.

This change is a bit of a clean up, but it is also required for the work
on improving fork testing (#1448). The improved fork testing will start
sourcing network data from a variety of sources but will need to cache
that data for subsequent test runs to ensure tests stay fast and
deterministic. The logic used to assign test names and numbers to envs
will be used to separate the caches for envs in tests. Even if we don't
end up using this in the fork testing flows I think this change stands
alone and is worth it for the other reasons above.

For #1448