Add safety notes about docs fields in soroban-spec-rust (#1650)
### What
Add and update comments in soroban-spec-rust explaining why the "docs"
fields from spec entries are intentionally not output as rustdocs in
generated Rust code.

### Why
Rustdocs can contain Rust code that gets executed. Generated code may
come from untrusted WASMs with untrusted spec docs, creating a code
execution risk.

We already have a comment about this in one of the files, but I wanted
to put this important note at the top of all the files in this crate so
it's present to anybody who's reading any of them.